refactor: MAT-269 선수 교체 WS 훅을 선수 교체 Adapter에서 직접 호출하게 하고, 기존 onSwap prop 제거
