+++ src/components/ToggleableStartingPlayers/ToggleableStartingPlayers.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
-import { MatchUserResponse, TeamResponse } from '@/apis/models';
+import type { MatchUserResponse, TeamResponse } from '@/apis/models';
+import { PlayerOnFieldGrid } from '@/components';
 import { PlayerList } from '@/components/PlayerList';
-import { PlayerOnFieldGrid } from '@/components/PlayerOnFieldGrid';
 import { type SubstitutionSourceType } from '@/stores';
 
 import * as styles from './ToggleableStartingPlayers.css';
@@ -11,14 +11,12 @@ interface ToggleableStartingPlayersProps {
   mode: SubstitutionSourceType;
   team: TeamResponse;
   players: MatchUserResponse[];
-  onSwap: (inPlayerId: number, outPlayerId: number) => void;
 }
 
 export const ToggleableStartingPlayers = ({
   mode,
   team,
   players,
-  onSwap,
 }: ToggleableStartingPlayersProps) => {
   const [isGridView, setIsGridView] = useState(true);
 
@@ -31,14 +29,9 @@ export const ToggleableStartingPlayers = ({
         {isGridView ? '리스트 보기' : '포메이션 보기'}
       </button>
       {isGridView ? (
-        <PlayerOnFieldGrid
-          mode={mode}
-          team={team}
-          players={players}
-          onSwap={onSwap}
-        />
+        <PlayerOnFieldGrid mode={mode} team={team} players={players} />
       ) : (
-        <PlayerList mode={mode} team={team} players={players} onSwap={onSwap} />
+        <PlayerList mode={mode} team={team} players={players} />
       )}
     </div>
   );
